# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "System API",
        "description": "System API",
        "version": "0.1"
    },
    "paths": {
        "/api/v1/categories": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "List Category",
                "operationId": "7ad1d64e6d8e9a6fb465f8b457d6d8ba",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/general--search"
                    },
                    {
                        "$ref": "#/components/parameters/general--searchFields"
                    },
                    {
                        "$ref": "#/components/parameters/general--searchJoin"
                    },
                    {
                        "$ref": "#/components/parameters/general--filter"
                    },
                    {
                        "$ref": "#/components/parameters/general--orderBy"
                    },
                    {
                        "$ref": "#/components/parameters/general--sortedBy"
                    },
                    {
                        "$ref": "#/components/parameters/general--with"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "total": {
                                            "type": "integer",
                                            "default": "15"
                                        },
                                        "has_more": {
                                            "type": "boolean",
                                            "default": "true"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/CategoryResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerToken": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Category"
                ],
                "summary": "Store Category",
                "operationId": "bbf601724507ed90b1a48d7dc478a9c7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CategoryRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CategoryResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerToken": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Category"
                ],
                "summary": "Store Category",
                "operationId": "837668b8e725006ae43bef019a2fe720",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CategoryRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CategoryResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerToken": []
                    }
                ]
            }
        },
        "/api/v1/categories/{id}": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "Show Category",
                "operationId": "9cb8a4648d805945ea10402b8237450b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "$ref": "#/components/parameters/general--with"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CategoryResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerToken": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Category"
                ],
                "summary": "Delete Category",
                "operationId": "5e8af896106f94516dc4cc0b791dec47",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerToken": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "CategoryRequest": {
                "title": "Category create request",
                "description": "Category create request body data",
                "required": [
                    "title",
                    "order",
                    "status"
                ],
                "properties": {
                    "title": {
                        "title": "title",
                        "type": "string"
                    },
                    "description": {
                        "title": "description",
                        "description": "maximum 255 charactor",
                        "type": "string"
                    },
                    "parent_id": {
                        "title": "parent_id",
                        "type": "integer"
                    },
                    "is_featured": {
                        "title": "is_featured",
                        "type": "boolean",
                        "example": "true"
                    },
                    "order": {
                        "title": "order",
                        "type": "integer",
                        "example": "1"
                    },
                    "status": {
                        "title": "status",
                        "type": "string",
                        "enum": [
                            "publish: Publish",
                            "un_publish: Un Publish",
                            "draft: Draft"
                        ],
                        "example": "publish"
                    }
                },
                "type": "object"
            },
            "CategoryResource": {
                "title": "CategoryResource",
                "description": "CategoryResource",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "title": {
                        "type": "string"
                    },
                    "parent_id": {
                        "type": "integer"
                    },
                    "description": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    },
                    "order": {
                        "type": "integer"
                    },
                    "is_featured": {
                        "type": "integer"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string",
                        "example": "2022-09-23T00:58:57.000000Z"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "CategoryResource"
                }
            },
...
        },
        "parameters": {
            "general--filter": {
                "name": "filter",
                "in": "query",
                "schema": {
                    "type": "string",
                    "default": "",
                    "example": "id,title"
                }
            },
            "general--orderBy": {
                "name": "orderBy",
                "in": "query",
                "schema": {
                    "type": "string",
                    "default": "",
                    "example": "id;created_at"
                }
            },
            "general--searchFields": {
                "name": "searchFields",
                "in": "query",
                "schema": {
                    "type": "string",
                    "default": "",
                    "example": "title:like;is_featured:=;status= || price:between || price:in"
                }
            },
            "general--searchJoin": {
                "name": "searchJoin",
                "in": "query",
                "description": "Default query or, if input and will query with AND",
                "schema": {
                    "type": "string",
                    "default": "or",
                    "example": "and"
                }
            },
            "general--search": {
                "name": "search",
                "in": "query",
                "schema": {
                    "type": "string",
                    "default": "",
                    "example": "title:Thai;is_featured:1;status:111 || price:300,500 || price:300,500"
                }
            },
            "general--sortedBy": {
                "name": "sortedBy",
                "in": "query",
                "schema": {
                    "type": "string",
                    "default": "",
                    "example": "desc"
                }
            },
            "general--with": {
                "name": "with",
                "in": "query",
                "schema": {
                    "type": "string",
                    "default": "",
                    "example": "createdBy;parent;updatedBy;posts"
                }
            }
        },
        "securitySchemes": {
            "bearerToken": {
                "type": "http",
                "name": "bearerToken",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}